# services/api/src/index.ts
import Fastify, { FastifyInstance } from "fastify";
import { Client } from "pg";
import { z } from "zod";
import Redis from "ioredis";
import { s3Artifacts } from "./s3-helpers.js";
import { SecretsManagerClient, GetSecretValueCommand } from "@aws-sdk/client-secrets-manager";
import client from "prom-client";

const app: FastifyInstance = Fastify({ logger: true });
// Resolve secrets from AWS Secrets Manager if configured
async function resolveSecret(name?: string) {
  if (!name) return undefined;
  try {
    const client = new SecretsManagerClient({});
    const res = await client.send(new GetSecretValueCommand({ SecretId: name }));
    return res.SecretString || undefined;
  } catch {
    return undefined;
  }
}

const dbUrl = (await resolveSecret(process.env.AWS_SECRETS_DB_URL)) || process.env.DB_URL;
const redisUrlEnv = (await resolveSecret(process.env.AWS_SECRETS_REDIS_URL)) || process.env.REDIS_URL || "redis://localhost:6379";

const pg = new Client({ connectionString: dbUrl });
await pg.connect().catch((err: unknown) => {
  app.log.error({ err }, "pg connect failed");
});
// Set tenant_id for RLS (example: from header)
app.addHook("onRequest", async (req) => {
  const tenantId = (req.headers["x-tenant-id"] || "t_demo").toString();
  try {
    await pg.query("select set_config('app.tenant_id', $1, true)", [tenantId]);
  } catch {}
});
const redis = new Redis(redisUrlEnv);

app.get("/health", async () => ({ ok: true }));
// --- Simple security middleware: IP allowlist + shared secret + PII redaction ---
function redactPIIString(s: string): string {
  return s
    .replace(/[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}/gi, "[redacted-email]")
    .replace(/(\+?\d[\d\s\-()]{7,})/g, "[redacted-phone]")
    .replace(/\b\d{4}\s?\d{4}\s?\d{4}\b/g, "[redacted-aadhaar]")
    .replace(/\b[A-Z]{5}\d{4}[A-Z]\b/g, "[redacted-pan]");
}
function redactObject<T>(obj: T): T {
  try {
    if (obj == null) return obj;
    if (typeof obj === "string") return redactPIIString(obj) as unknown as T;
    if (Array.isArray(obj)) return obj.map((v) => redactObject(v)) as unknown as T;
    if (typeof obj === "object") {
      const out: any = Array.isArray(obj) ? [] : {};
      for (const [k, v] of Object.entries(obj as any)) {
        const key = k.toLowerCase();
        if (["email", "phone", "mobile", "aadhaar", "pan", "ssn", "card", "credit", "token"].some((x) => key.includes(x))) {
          out[k] = "[redacted]";
        } else if (typeof v === "string") {
          out[k] = redactPIIString(v);
        } else {
          out[k] = redactObject(v);
        }
      }
      return out as T;
    }
    return obj;
  } catch {
    return obj;
  }
}

app.addHook("onRequest", async (req, reply) => {
  const allowed = (process.env.API_ALLOWED_IPS || "").split(",").map((s) => s.trim()).filter(Boolean);
  if (allowed.length > 0 && !allowed.includes(req.ip)) {
    req.log.warn({ ip: req.ip }, "blocked ip");
    return reply.code(403).send({ code: "forbidden" });
  }
  const shared = process.env.API_SHARED_SECRET;
  if (shared) {
    const token = (req.headers["x-internal-token"] || "").toString();
    if (token !== shared) {
      return reply.code(401).send({ code: "unauthorized" });
    }
  }
  // Sanitize incoming payloads
  (req as any).body = redactObject((req as any).body);
  (req as any).query = redactObject((req as any).query);
  (req as any).params = redactObject((req as any).params);
});

// Prometheus metrics endpoint
const registry = new client.Registry();
client.collectDefaultMetrics({ register: registry });
const httpRequests = new client.Counter({
  name: "http_requests_total",
  help: "Total HTTP requests",
  labelNames: ["method", "path", "status"],
});
registry.registerMetric(httpRequests);

app.addHook("onResponse", async (req, reply) => {
  try {
    httpRequests.inc({ method: req.method, path: req.routerPath || req.url, status: String(reply.statusCode) });
  } catch {}
});

app.get("/metrics", async (req, reply) => {
  reply.header("content-type", registry.contentType);
  return await registry.metrics();
});

const createAgentSchema = z.object({
  name: z.string(),
  config: z.record(z.any()),
});

app.post("/v1/agents", async (req, reply) => {
  const parsed = createAgentSchema.safeParse((req as any).body);
  if (!parsed.success) {
    return reply.code(400).send({ code: "bad_request", error: parsed.error.flatten() });
  }
  const id = `a_${Math.random().toString(36).slice(2)}`;
  try {
    await pg.query("insert into agents (id, tenant_id, name, config) values ($1, $2, $3, $4)",
      [id, "t_demo", parsed.data.name, JSON.stringify(parsed.data.config)]);
  } catch (err) {
    app.log.error({ err }, "insert agent failed");
  }
  return { id, version: 1 };
});

const createCallSchema = z.object({
  agentId: z.string(),
  to: z.string(),
  from: z.string().optional(),
  metadata: z.record(z.any()).optional(),
});

app.post("/v1/calls", async (req, reply) => {
  const parsed = createCallSchema.safeParse((req as any).body);
  if (!parsed.success) {
    return reply.code(400).send({ code: "bad_request", error: parsed.error.flatten() });
  }
  
  // Check tenant caps
  const tenantId = (req.headers["x-tenant-id"] || "t_demo").toString();
  
  // Check concurrent calls limit
  const concurrentCalls = await pg.query(
    "select count(*) from calls where tenant_id = $1 and status in ('created', 'active') and started_at > now() - interval '1 hour'",
    [tenantId]
  );
  
  const callCount = parseInt(concurrentCalls.rows[0]?.count || "0");
  const maxConcurrent = parseInt(process.env.MAX_CONCURRENT_CALLS || "10");
  
  if (callCount >= maxConcurrent) {
    return reply.code(429).send({
      code: "rate_limit_exceeded",
      message: "Maximum concurrent calls reached"
    });
  }
  
  // Check daily usage cap
  const dailyUsage = await pg.query(
    "select sum(cost_inr) as total from calls where tenant_id = $1 and started_at > now() - interval '24 hours'",
    [tenantId]
  );
  
  const dailyTotal = parseFloat(dailyUsage.rows[0]?.total || "0");
  const dailyCap = parseFloat(process.env.DAILY_COST_CAP_INR || "10000");
  
  if (dailyTotal >= dailyCap) {
    return reply.code(429).send({
      code: "usage_cap_exceeded",
      message: "Daily usage cap exceeded"
    });
  }
  
  const id = `c_${Math.random().toString(36).slice(2)}`;
  try {
    await pg.query(
      "insert into calls (id, tenant_id, agent_id, direction, from_num, to_num, status, started_at) values ($1, $2, $3, $4, $5, $6, $7, now())",
      [id, tenantId, parsed.data.agentId, "outbound", parsed.data.from || "system", parsed.data.to, "created"]
    );
    
    // Store metadata if provided
    if (parsed.data.metadata) {
      await redis.hset(`call:${id}:metadata`, parsed.data.metadata);
    }
  } catch (err) {
    app.log.error({ err }, "insert call failed");
    return reply.code(500).send({ code: "internal_error" });
  }
  
    // If outbound dialing is enabled, request Jambonz to originate a call
    const jambonzUrl = process.env.JAMBONZ_OUTCALL_URL; // e.g., https://<jambonz>/v1/Accounts/<sid>/Calls
    const jambonzToken = process.env.JAMBONZ_TOKEN;     // Basic or Bearer depending on setup
    const telephonyHook = process.env.TELEPHONY_CALL_HOOK || "http://telephony:8085/call";
    if (jambonzUrl && jambonzToken) {
      try {
        const jb = await fetch(jambonzUrl, {
          method: "POST",
          headers: {
            "content-type": "application/json",
            "authorization": jambonzToken,
          },
          body: JSON.stringify({
            from: parsed.data.from || "sbc",
            to: parsed.data.to,
            application_sid: process.env.JAMBONZ_APP_SID,
            call_hook: { url: telephonyHook, method: "POST" },
            call_status_hook: { url: `${telephonyHook.replace(/\/call$/, "")}/status/${id}`, method: "POST" },
            timeout: 30,
          }),
        });
        if (!jb.ok) {
          app.log.warn({ status: jb.status }, "jambonz originate failed");
        }
      } catch (err) {
        app.log.warn({ err }, "jambonz originate error");
      }
    }

  return { id, status: jambonzUrl ? "dialing" : "created" };
});

app.get("/v1/calls/:id/timeline", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  try {
    const entries = await redis.xrange(`events:${id}`, "-", "+", "COUNT", 500);
    const timeline = entries.map(([, fields]) => {
      const rec: Record<string, string> = {} as any;
      for (let i = 0; i < fields.length; i += 2) rec[String(fields[i])] = String(fields[i + 1]);
      return { kind: rec.kind, payload: safeJson(rec.payload) };
    });
    return { callId: id, timeline };
  } catch (err) {
    app.log.error({ err }, "timeline error");
    return reply.code(500).send({ code: "error" });
  }
});

// Get signed URLs for artifacts
app.get("/v1/calls/:id/artifacts", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  
  try {
    const artifacts = await s3Artifacts.getCallArtifacts(id);
    return artifacts;
  } catch (err) {
    app.log.error({ err }, "Failed to get artifacts");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Upload call recording
app.post("/v1/calls/:id/recording", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  const body = (req as any).body;
  
  if (!body || !Buffer.isBuffer(body)) {
    return reply.code(400).send({ code: "bad_request", message: "Body must be audio data" });
  }
  
  try {
    const s3Path = await s3Artifacts.uploadRecording(id, body);
    
    // Update call record
    await pg.query(
      "update calls set status = 'completed', ended_at = now() where id = $1",
      [id]
    );
    
    return { success: true, path: s3Path };
  } catch (err) {
    app.log.error({ err }, "Failed to upload recording");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Get call details with costs
app.get("/v1/calls/:id", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  
  try {
    const callResult = await pg.query(
      "select * from calls where id = $1",
      [id]
    );
    
    if (callResult.rows.length === 0) {
      return reply.code(404).send({ code: "not_found" });
    }
    
    const call = callResult.rows[0];
    
    // Get costs breakdown
    const costsResult = await pg.query(
      "select * from call_costs where call_id = $1",
      [id]
    );
    
    // Get metadata from Redis
    const metadata = await redis.hgetall(`call:${id}:metadata`);
    
    return {
      ...call,
      costs: costsResult.rows,
      metadata: Object.keys(metadata).length > 0 ? metadata : undefined,
    };
  } catch (err) {
    app.log.error({ err }, "Failed to get call details");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Submit call summary
app.post("/v1/calls/:id/summary", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  const summary = (req as any).body;
  
  try {
    // Upload summary to S3
    const s3Path = await s3Artifacts.uploadSummary(id, summary);
    
    // Trigger summary webhook
    const tenantWebhook = process.env.TENANT_WEBHOOK_URL;
    if (tenantWebhook) {
      await redis.lpush(
        "webhooks:queue",
        JSON.stringify({
          id: `wh_${Date.now()}`,
          url: tenantWebhook,
          headers: { "content-type": "application/json" },
          body: JSON.stringify({
            type: "call.summary",
            callId: id,
            summary,
            artifactUrl: s3Path,
          }),
          attempts: 0,
        })
      );
    }
    
    return { success: true, path: s3Path };
  } catch (err) {
    app.log.error({ err }, "Failed to submit summary");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// List all calls with pagination and filtering
app.get("/v1/calls", async (req, reply) => {
  const query = (req.query as any) as {
    page?: string;
    limit?: string;
    status?: string;
    agentId?: string;
    from?: string;
    to?: string;
    tenantId?: string;
  };
  
  const page = parseInt(query.page || "1");
  const limit = Math.min(parseInt(query.limit || "50"), 100);
  const offset = (page - 1) * limit;
  
  try {
    let whereClause = "where 1=1";
    const params: any[] = [];
    let paramIndex = 1;
    
    if (query.status) {
      whereClause += ` and status = $${paramIndex++}`;
      params.push(query.status);
    }
    
    if (query.agentId) {
      whereClause += ` and agent_id = $${paramIndex++}`;
      params.push(query.agentId);
    }
    
    if (query.from) {
      whereClause += ` and started_at >= $${paramIndex++}`;
      params.push(query.from);
    }
    
    if (query.to) {
      whereClause += ` and started_at <= $${paramIndex++}`;
      params.push(query.to);
    }
    
    const tenantId = query.tenantId || (req.headers["x-tenant-id"] || "t_demo").toString();
    whereClause += ` and tenant_id = $${paramIndex++}`;
    params.push(tenantId);
    
    // Get total count
    const countResult = await pg.query(
      `select count(*) from calls ${whereClause}`,
      params
    );
    const total = parseInt(countResult.rows[0]?.count || "0");
    
    // Get calls with pagination
    const callsResult = await pg.query(
      `select * from calls ${whereClause} order by started_at desc limit $${paramIndex++} offset $${paramIndex++}`,
      [...params, limit, offset]
    );
    
    return {
      calls: callsResult.rows,
      pagination: {
        page,
        limit,
        total,
        pages: Math.ceil(total / limit),
      },
    };
  } catch (err) {
    app.log.error({ err }, "Failed to list calls");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Update call status
app.patch("/v1/calls/:id/status", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  const { status, metadata } = (req.body as any) || {};
  
  if (!status) {
    return reply.code(400).send({ code: "bad_request", message: "status required" });
  }
  
  try {
    const updateResult = await pg.query(
      "update calls set status = $1, updated_at = now() where id = $2 returning *",
      [status, id]
    );
    
    if (updateResult.rows.length === 0) {
      return reply.code(404).send({ code: "not_found" });
    }
    
    // Update metadata if provided
    if (metadata) {
      await redis.hset(`call:${id}:metadata`, metadata);
    }
    
    // Publish status change to timeline
    await redis.xadd(
      `events:${id}`,
      "*",
      "kind",
      "call.status_changed",
      "payload",
      JSON.stringify({ status, metadata })
    );
    
    return { success: true, call: updateResult.rows[0] };
  } catch (err) {
    app.log.error({ err }, "Failed to update call status");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Get agent details
app.get("/v1/agents/:id", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  
  try {
    const agentResult = await pg.query(
      "select * from agents where id = $1",
      [id]
    );
    
    if (agentResult.rows.length === 0) {
      return reply.code(404).send({ code: "not_found" });
    }
    
    // Get agent statistics
    const statsResult = await pg.query(
      "select count(*) as total_calls, avg(cost_inr) as avg_cost from calls where agent_id = $1",
      [id]
    );
    
    const agent = agentResult.rows[0];
    const stats = statsResult.rows[0];
    
    return {
      ...agent,
      stats: {
        totalCalls: parseInt(stats.total_calls || "0"),
        averageCost: parseFloat(stats.avg_cost || "0"),
      },
    };
  } catch (err) {
    app.log.error({ err }, "Failed to get agent details");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// List all agents
app.get("/v1/agents", async (req, reply) => {
  const query = (req.query as any) as {
    page?: string;
    limit?: string;
    tenantId?: string;
  };
  
  const page = parseInt(query.page || "1");
  const limit = Math.min(parseInt(query.limit || "50"), 100);
  const offset = (page - 1) * limit;
  
  try {
    const tenantId = query.tenantId || (req.headers["x-tenant-id"] || "t_demo").toString();
    
    // Get total count
    const countResult = await pg.query(
      "select count(*) from agents where tenant_id = $1",
      [tenantId]
    );
    const total = parseInt(countResult.rows[0]?.count || "0");
    
    // Get agents with pagination
    const agentsResult = await pg.query(
      "select * from agents where tenant_id = $1 order by created_at desc limit $2 offset $3",
      [tenantId, limit, offset]
    );
    
    return {
      agents: agentsResult.rows,
      pagination: {
        page,
        limit,
        total,
        pages: Math.ceil(total / limit),
      },
    };
  } catch (err) {
    app.log.error({ err }, "Failed to list agents");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Update agent configuration
app.patch("/v1/agents/:id", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  const { name, config, status } = (req.body as any) || {};
  
  try {
    const updateFields: string[] = [];
    const params: any[] = [];
    let paramIndex = 1;
    
    if (name !== undefined) {
      updateFields.push(`name = $${paramIndex++}`);
      params.push(name);
    }
    
    if (config !== undefined) {
      updateFields.push(`config = $${paramIndex++}`);
      params.push(JSON.stringify(config));
    }
    
    if (status !== undefined) {
      updateFields.push(`status = $${paramIndex++}`);
      params.push(status);
    }
    
    if (updateFields.length === 0) {
      return reply.code(400).send({ code: "bad_request", message: "No fields to update" });
    }
    
    updateFields.push(`updated_at = now()`);
    
    const updateResult = await pg.query(
      `update agents set ${updateFields.join(", ")} where id = $${paramIndex++} returning *`,
      [...params, id]
    );
    
    if (updateResult.rows.length === 0) {
      return reply.code(404).send({ code: "not_found" });
    }
    
    return { success: true, agent: updateResult.rows[0] };
  } catch (err) {
    app.log.error({ err }, "Failed to update agent");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Delete agent
app.delete("/v1/agents/:id", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  
  try {
    // Check if agent has active calls
    const activeCallsResult = await pg.query(
      "select count(*) from calls where agent_id = $1 and status in ('created', 'active')",
      [id]
    );
    
    const activeCalls = parseInt(activeCallsResult.rows[0]?.count || "0");
    if (activeCalls > 0) {
      return reply.code(400).send({
        code: "agent_in_use",
        message: "Cannot delete agent with active calls",
        activeCalls,
      });
    }
    
    const deleteResult = await pg.query(
      "delete from agents where id = $1 returning *",
      [id]
    );
    
    if (deleteResult.rows.length === 0) {
      return reply.code(404).send({ code: "not_found" });
    }
    
    return { success: true, deleted: deleteResult.rows[0] };
  } catch (err) {
    app.log.error({ err }, "Failed to delete agent");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Get tenant usage statistics
app.get("/v1/tenants/:id/usage", async (req, reply) => {
  const { id } = (req.params as any) as { id: string };
  const { period } = (req.query as any) as { period?: string };
  
  const periodMap: Record<string, string> = {
    "1h": "1 hour",
    "24h": "24 hours",
    "7d": "7 days",
    "30d": "30 days",
    "1m": "1 month",
  };
  
  const timeRange = periodMap[period || "24h"] || "24 hours";
  
  try {
    // Get call statistics
    const callStatsResult = await pg.query(
      `select 
        count(*) as total_calls,
        count(case when status = 'completed' then 1 end) as completed_calls,
        count(case when status = 'failed' then 1 end) as failed_calls,
        avg(cost_inr) as avg_cost,
        sum(cost_inr) as total_cost,
        avg(extract(epoch from (ended_at - started_at))) as avg_duration
      from calls 
      where tenant_id = $1 and started_at > now() - interval '${timeRange}'`,
      [id]
    );
    
    const stats = callStatsResult.rows[0];
    
    // Get agent statistics
    const agentStatsResult = await pg.query(
      "select count(*) as total_agents, count(case when status = 'active' then 1 end) as active_agents from agents where tenant_id = $1",
      [id]
    );
    
    const agentStats = agentStatsResult.rows[0];
    
    return {
      tenantId: id,
      period,
      timeRange,
      calls: {
        total: parseInt(stats.total_calls || "0"),
        completed: parseInt(stats.completed_calls || "0"),
        failed: parseInt(stats.failed_calls || "0"),
        averageCost: parseFloat(stats.avg_cost || "0"),
        totalCost: parseFloat(stats.total_cost || "0"),
        averageDuration: parseFloat(stats.avg_duration || "0"),
      },
      agents: {
        total: parseInt(agentStats.total_agents || "0"),
        active: parseInt(agentStats.active_agents || "0"),
      },
    };
  } catch (err) {
    app.log.error({ err }, "Failed to get tenant usage");
    return reply.code(500).send({ code: "internal_error" });
  }
});

// Health check with database connectivity
app.get("/health/detailed", async () => {
  try {
    // Check database connectivity
    await pg.query("select 1");
    
    // Check Redis connectivity
    await redis.ping();
    
    return {
      ok: true,
      database: "connected",
      redis: "connected",
      timestamp: new Date().toISOString(),
    };
  } catch (err) {
    app.log.error({ err }, "Health check failed");
    return {
      ok: false,
      error: (err as Error).message,
      timestamp: new Date().toISOString(),
    };
  }
});

function safeJson(v?: string) {
  try { return v ? JSON.parse(v) : null; } catch { return v ?? null; }
}

const PORT = Number(process.env.PORT || 8080);

app.listen({ port: PORT, host: "0.0.0.0" }).catch((err) => {
  app.log.error(err);
  process.exit(1);
});